# Replace PrefixLogger with LogfmtLogger (fix for #34).