--- a/addsvc/main.go
+++ b/addsvc/main.go
@@ -48,7 +48,7 @@ func main() {
 
 	// `package log` domain
 	var logger kitlog.Logger
-	logger = kitlog.NewPrefixLogger(os.Stderr)
+	logger = kitlog.NewLogfmtLogger(os.Stderr)
 	logger = kitlog.With(logger, "ts", kitlog.DefaultTimestampUTC)
 	kitlog.DefaultLogger = logger                     // for other gokit components
 	stdlog.SetOutput(kitlog.NewStdlibAdapter(logger)) // redirect stdlib logging to us