--- a/log/default_logger.go
+++ b/log/default_logger.go
@@ -1,5 +1,5 @@
 package log
 
-// DefaultLogger is used by gokit components. By default, it's a PrefixLogger
+// DefaultLogger is used by gokit components. By default, it's a LogfmtLogger
 // that writes to the stdlib log.
-var DefaultLogger = NewPrefixLogger(StdlibWriter{})
+var DefaultLogger = NewLogfmtLogger(StdlibWriter{})