--- a/log/logfmt_logger.go
+++ b/log/logfmt_logger.go
@@ -11,12 +11,17 @@ type logfmtLogger struct {
 }
 
 // NewLogfmtLogger returns a logger that encodes keyvals to the Writer in
-// logfmt format.
+// logfmt format. The passed Writer must be safe for concurrent use by
+// multiple goroutines if the returned Logger will be used concurrently.
 func NewLogfmtLogger(w io.Writer) Logger {
 	return &logfmtLogger{w}
 }
 
 func (l logfmtLogger) Log(keyvals ...interface{}) error {
+	// The Logger interface requires implementations to be safe for concurrent
+	// use by multiple goroutines. For this implementation that means making
+	// only one call to l.w.Write() for each call to Log. We first collect all
+	// of the bytes into b, and then call l.w.Write(b).
 	b, err := logfmt.MarshalKeyvals(keyvals...)
 	if err != nil {
 		return err