- Fix for #34.
- Adds dependency on gopkg.in/logfmt.v0.
- Breaks logging of compound data types until the logfmt package has support for that.
- Improves performance:

Before:
```
BenchmarkPrefixLoggerSimple       500000      2916 ns/op     322 B/op          8 allocs/op
BenchmarkPrefixLoggerContextual   300000      4723 ns/op     531 B/op         13 allocs/op
```
After:
```
BenchmarkLogfmtLoggerSimple      1000000      2155 ns/op     272 B/op          6 allocs/op
BenchmarkLogfmtLoggerContextual   500000      3502 ns/op     480 B/op         10 allocs/op
```