@ChrisHines 2015-05-26
See https://github.com/go-logfmt/logfmt/issues/1 for discussion about compound data types.