@ChrisHines 2015-05-30
I just pushed https://github.com/go-logfmt/logfmt/commit/c50dc9aa845e681529d8e49d3dc11ec0195aacac which I believe makes go-logfmt/logfmt good enough to merge this PR.