@peterbourgon 2015-05-30
Those points aside, LGTM.